pyproject.toml
[project]
name = "python-parser"
version = "0.1.0"
description = ""
authors = [
    {name = "iiTzJambo",email = "morzhmorzh46@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4 (==4.13.4)",
    "translate (>=3.6.1,<4.0.0)",
    "fake-useragent (>=2.2.0,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
